- found a nice #pragma for MSVC that makes the FraggleScript initialization code 8kb shorter. - fixed: Sorting of the draw lists by texture did not take clamping into account. Changing   this gave another mild performance gain. - Made GLWall::SetWallCoordinates floating point only. - fixed: FGLRenderer::DrawBlend was missing a NULL pointer check for the player actor.
git-svn-id: http://mancubus.net/svn/hosted/gzdoom/trunk@612 b0f79afe-0144-0410-b225-9a4edf0717df

+++ src/gl/scene/gl_wall.h
@@ -175,7 +175,7 @@ private:
 	bool DoHorizon(seg_t * seg,sector_t * fs, vertex_t * v1,vertex_t * v2);
 
 	bool SetWallCoordinates(seg_t * seg, float ceilingrefheight,
-							int topleft,int topright, int bottomleft,int bottomright, int texoffset);
+							float topleft,float topright,float bottomleft,float bottomright, float texoffset);
 
 	void DoTexture(int type,seg_t * seg,int peg,
 						   int ceilingrefheight,int floorrefheight,
